+++ RELEASENOTES.md
@@ -7,6 +7,9 @@ These are the release notes for the [mediawiki-datamodel](README.md).
 * Revision objects now require a PageIdentifier object instead of a $pageId int
 * Page objects now require a PageIdentifier objects instead of a $title and $pageId
 * Page objects can be constructed without a Revisions object
+* Content getNativeData renamed to getData
+* Content constructor no longer takes the modelid
+* Content has new method getModel in places of random constants
 
 
 ## Version 0.4 (2014-07-08)